+++ src/events/RegisterLiveStatusesEvent.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\blitz\events;
+
+use yii\base\Event;
+
+class RegisterLiveStatusesEvent extends Event
+{
+    // Properties
+    // =========================================================================
+
+    /**
+     * @var string[]
+     */
+    public $liveStatuses = [];
+}